Suppress harmless diagnostics for test projects
FakeItEasy must work with a variety of client code that doesn't conform to
some coding "best practices". As such, our test need to include types that
likewise violate these rules. Including suppression statements for the
more common of the rules is tiresome and clutters the code, so move some
out into the global ruleset:

- CA1021: AvoidOutParameters
- CA1045: DoNotPassTypesByReference
- CA1062: ValidateArgumentsOfPublicMethods
- CA1716: IdentifiersShouldNotMatchKeywords